wz/backend/requirements.txt
Flask==2.0.3
python-dotenv==0.20.0
Flask-Cors==4.0.0
requests==2.27.1
